python_server/requirements.txt
flask
flask-cors
python-dotenv
firebase-admin